merges qt4 branches back to trunk
git-svn-id: http://svn.mythtv.org/svn/trunk@16790 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythui/mythgesture.cpp
@@ -272,7 +272,7 @@ QString MythGesture::translate(void)
 bool MythGesture::record(const QPoint & p)
 {
     /* only record if we haven't exceeded the maximum points */
-    if ((points.size() >= max_points) || !recording())
+    if (((uint)points.size() >= max_points) || !recording())
         return false;
 
     if (points.size() == 0)
